+++ src/main/java/com/oitsjustjose/vtweaks/util/ConfigItemParser.java
@@ -2,20 +2,22 @@ package com.oitsjustjose.vtweaks.util;
 
 import java.util.ArrayList;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 import com.oitsjustjose.vtweaks.VTweaks;
 
-import net.minecraft.block.Block;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
 
 public class ConfigItemParser
 {
+	private static Logger LOGGER = LogManager.getLogger(VTweaks.MODID);
+
 	public static void parseItems()
 	{
 		ArrayList<ItemStack> stackList = new ArrayList<ItemStack>();
 
-		LogHelper.info(">> Running Config Item Parser");
+		LOGGER.info(">> Running Config Item Parser");
 
 		for (String s : VTweaks.config.challengerMobLoot)
 		{
@@ -26,7 +28,7 @@ public class ConfigItemParser
 
 			if (parts.length == 4)
 			{
-				ItemStack temp = findItemStack(parts[0], parts[1]);
+				ItemStack temp = HelperFunctions.findItemStack(parts[0], parts[1]);
 
 				if (temp != null)
 				{
@@ -40,13 +42,13 @@ public class ConfigItemParser
 					}
 					catch (NumberFormatException e)
 					{
-						LogHelper.info("There was a number formatting issue with entry: " + s + ". It has been skipped.");
+						LOGGER.info("There was a number formatting issue with entry: " + s + ". It has been skipped.");
 					}
 				}
 			}
 			else if (parts.length == 3)
 			{
-				ItemStack temp = findItemStack(parts[0], parts[1]);
+				ItemStack temp = HelperFunctions.findItemStack(parts[0], parts[1]);
 
 				if (temp != null)
 				{
@@ -58,14 +60,14 @@ public class ConfigItemParser
 					}
 					catch (NumberFormatException e)
 					{
-						LogHelper.info("There was a number formatting issue with entry: " + s + ". It has been skipped.");
+						LOGGER.info("There was a number formatting issue with entry: " + s + ". It has been skipped.");
 					}
-					
+
 				}
 			}
 			else if (parts.length == 2)
 			{
-				ItemStack temp = findItemStack(parts[0], parts[1]);
+				ItemStack temp = HelperFunctions.findItemStack(parts[0], parts[1]);
 
 				if (temp != null)
 				{
@@ -75,23 +77,11 @@ public class ConfigItemParser
 			// The string was not formatted correctly. User is notified.
 			else
 			{
-				LogHelper.info("There was an error parsing item " + s + " as a challenger mob drop. Please confirm that your formatting is correct.");
+				LOGGER.info("There was an error parsing item " + s + " as a challenger mob drop. Please confirm that your formatting is correct.");
 			}
 		}
 
-		LogHelper.info(">> Config Item Parsing complete!");
+		LOGGER.info(">> Config Item Parsing complete!");
 		VTweaks.config.setChallengerLootTable(stackList);
-	}
-
-	public static ItemStack findItemStack(String modid, String name)
-	{
-		ResourceLocation resLoc = new ResourceLocation(modid, name);
-		if (Item.REGISTRY.containsKey(resLoc))
-			return new ItemStack(Item.REGISTRY.getObject(resLoc));
-		else if (Block.REGISTRY.containsKey(resLoc))
-			if (Item.getItemFromBlock(Block.REGISTRY.getObject(resLoc)) != null)
-				return new ItemStack(Item.getItemFromBlock(Block.REGISTRY.getObject(resLoc)), 1);
-
-		return null;
 	}
 }